refactor(frontend): reorganize file structure
